+++ admin/content/tambah-reviewer.php
@@ -6,18 +6,25 @@
         $rating = $_POST["rating"];
         $quote = $_POST["quote"];
         $photo = $_FILES["photo"]["name"];
-        $tmp_name = $_FILES["photo"]["tmp_name"];
-        $filename = uniqid() . "_" . basename($photo);
-        $filepath = "uploads/" . $filename;
-        unlink("uploads/" . $filename);
-        move_uploaded_file($tmp_name, $filepath);
-
-        $ekstensi = ["png", "jpg", "jpeg"];
-        $eks = pathinfo($photo, PATHINFO_EXTENSION);
-        if (!in_array($eks, $ekstensi)) {
-            $error[] = "Mohon maaf, ekstensi file tidak ditemukan";
-        }else{
-            $query = mysqli_query($config, "INSERT INTO testimoni(photo, name, profesion, rating, quote) VALUES('$filename', '$name', '$profesi', '$rating', '$quote')");
+        if (!empty($photo)) {
+            $tmp_name = $_FILES["photo"]["tmp_name"];
+            $filename = uniqid() . "_" . basename($photo);
+            $filepath = "uploads/" . $filename;
+            unlink("uploads/" . $filename);
+            move_uploaded_file($tmp_name, $filepath);
+    
+            $ekstensi = ["png", "jpg", "jpeg"];
+            $eks = pathinfo($photo, PATHINFO_EXTENSION);
+            if (!in_array($eks, $ekstensi)) {
+                $error[] = "Mohon maaf, ekstensi file tidak ditemukan";
+            }else{
+                $query = mysqli_query($config, "INSERT INTO testimoni(photo, name, profesion, rating, quote) VALUES('$filename', '$name', '$profesi', '$rating', '$quote')");
+                if ($query) {
+                    header("location:?page=testimonials&tambah=berhasil");
+                }
+            }
+        }else {
+            $query = mysqli_query($config, "INSERT INTO testimoni(name, profesion, rating, quote) VALUES('$name', '$profesi', '$rating', '$quote')");
             if ($query) {
                 header("location:?page=testimonials&tambah=berhasil");
             }
@@ -48,14 +55,21 @@
         $rating = $_POST["rating"];
         $quote = $_POST["quote"];
         $photo = $_FILES["photo"]["name"];
-        $tmp_name = $_FILES["photo"]["tmp_name"];
-        $filename = uniqid() . "_" . basename($photo);
-        $filepath = "uploads/" . $filename;
-        unlink("uploads/" . $filename);
-        move_uploaded_file($tmp_name, $filepath);
-        $queryUpdate = mysqli_query($config, "UPDATE testimoni SET photo='$filename', name='$name', profesion='$profesi', rating='$rating', quote='$quote' WHERE id='$id_user'");
-        if ($queryUpdate) {
-            header("location:?page=testimonials&ubah=berhasil");
+        if (!empty($photo)) {
+            $tmp_name = $_FILES["photo"]["tmp_name"];
+            $filename = uniqid() . "_" . basename($photo);
+            $filepath = "uploads/" . $filename;
+            unlink("uploads/" . $filename);
+            move_uploaded_file($tmp_name, $filepath);
+            $queryUpdate = mysqli_query($config, "UPDATE testimoni SET photo='$filename', name='$name', profesion='$profesi', rating='$rating', quote='$quote' WHERE id='$id_user'");
+            if ($queryUpdate) {
+                header("location:?page=testimonials&ubah=berhasil");
+            }
+        }else {
+            $queryUpdate = mysqli_query($config, "UPDATE testimoni SET name='$name', profesion='$profesi', rating='$rating', quote='$quote' WHERE id='$id_user'");
+            if ($queryUpdate) {
+                header("location:?page=testimonials&ubah=berhasil");
+            }
         }
     }
 ?>
@@ -67,11 +81,15 @@
         </div>
         <div class="col-sm-10">
             <input type="file" name="photo">
+            <?php if (isset($_GET["edit"])): ?>
+                <br><br>
+                <img src="uploads/<?php echo $rowEdit["photo"]; ?>" alt="" width="100">
+            <?php endif; ?>
         </div>
     </div>
     <div class="mb-3 row">
         <div class="col-sm-2">
-            <label for="">Nama *</label>
+            <label for="">Nama <span class="text-danger">*</span></label>
         </div>
         <div class="col-sm-10">
             <input required type="text" name="name" class="form-control" placeholder="Masukkan Nama Anda" value="<?php echo $nilaiName; ?>">
@@ -79,7 +97,7 @@
     </div>
     <div class="mb-3 row">
         <div class="col-sm-2">
-            <label for="">Jabatan *</label>
+            <label for="">Jabatan <span class="text-danger">*</span></label>
         </div>
         <div class="col-sm-10">
             <input required type="text" name="profesion" class="form-control" placeholder="Masukkan Jabatan Anda" value="<?php echo $nilaiProfesion; ?>">
@@ -87,7 +105,7 @@
     </div>
     <div class="mb-3 row">
         <div class="col-sm-2">
-            <label for="">Rating <?php echo isset($_GET["edit"]) ? "" : "*"; ?></label>
+            <label for="">Rating <span class="text-danger">*</span></label>
         </div>
         <div class="col-sm-10">
             <input required type="number" name="rating" min="0" max="5" class="form-control" placeholder="Masukkan Rating Anda" value="<?php echo $nilaiRating; ?>">
@@ -95,10 +113,10 @@
     </div>
     <div class="mb-3 row">
         <div class="col-sm-2">
-            <label for="">Quote</label>
+            <label for="">Quote <span class="text-danger">*</span></label>
         </div>
         <div class="col-sm-10">
-            <textarea name="quote" cols="30" rows="10" class="form-control"><?php echo $nilaiQuote; ?></textarea>
+            <textarea required id="summernote" name="quote" cols="30" rows="10" class="form-control"><?php echo $nilaiQuote; ?></textarea>
         </div>
     </div>
     <div class="mb-3 row" align="center">
